+++ src/Leges/Foundation/Plugin.php
@@ -23,7 +23,7 @@ class Plugin extends BasePlugin
      * Version of the plugin.
      * Used for setting versions of enqueue scripts and styles.
      */
-    const VERSION = '1.2.5';
+    const VERSION = '2.0.0';
 
     protected function checkForUpdate()
     {